feat: remove legacy player achievements
Refactor remaingin Awarded table usage.
