+++ app/Helpers/database/player-game.php
@@ -1,9 +1,6 @@
 <?php
 
 use App\Platform\Enums\AchievementFlag;
-use App\Platform\Enums\AchievementType;
-use App\Platform\Enums\UnlockMode;
-use App\Platform\Models\Achievement;
 use App\Platform\Models\Game;
 use App\Platform\Models\PlayerGame;
 use App\Platform\Models\PlayerSession;
@@ -12,77 +9,6 @@ use App\Site\Models\User;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\DB;
-
-/**
- * @deprecated TODO read from PlayerGame model, badge eligibility checks moved to UpdatePlayerGameBadgeMetrics
- */
-function testBeatenGame(int $gameId, string $user): array
-{
-    // First, get the count of beaten-tier achievements for the game.
-    // We'll use this to determine if the game is even beatable, and later
-    // use it to determine if the user has in fact beaten the game.
-    $gameTierAchievementCounts = Achievement::where('GameID', $gameId)
-        ->whereIn('type', [AchievementType::Progression, AchievementType::WinCondition])
-        ->where('Flags', AchievementFlag::OfficialCore)
-        ->select(['type', DB::raw('count(*) as total')])
-        ->groupBy('type')
-        ->get()
-        ->keyBy('type')
-        ->transform(function ($item) {
-            return $item->total;
-        });
-
-    $totalProgressions = (int) ($gameTierAchievementCounts[AchievementType::Progression] ?? 0);
-    $totalWinConditions = (int) ($gameTierAchievementCounts[AchievementType::WinCondition] ?? 0);
-
-    // If the game has no beaten-tier achievements assigned, it is not considered beatable.
-    // Bail.
-    if ($totalProgressions === 0 && $totalWinConditions === 0) {
-        // TODO use $playerGame->achievements_beat for isBeatable, remove rest
-        return [
-            'isBeatenSoftcore' => false,
-            'isBeatenHardcore' => false,
-            'isBeatable' => false,
-        ];
-    }
-
-    // We can now start checking if the user has beaten the game.
-    // Start by querying for their unlocked beaten-tier achievements.
-    $userAchievements = Achievement::where('Achievements.GameID', $gameId)
-        ->whereIn('Achievements.type', [AchievementType::Progression, AchievementType::WinCondition])
-        ->where('Achievements.Flags', AchievementFlag::OfficialCore)
-        ->leftJoin('Awarded', function ($join) use ($user) {
-            $join->on('Achievements.ID', '=', 'Awarded.AchievementID')
-                ->where('Awarded.User', '=', $user);
-        })
-        ->addSelect(['Achievements.type', 'Awarded.HardcoreMode', 'Awarded.AchievementID', 'Awarded.Date'])
-        ->orderByDesc('Awarded.Date')
-        ->get(['Achievements.type', 'Awarded.HardcoreMode', 'Awarded.AchievementID', 'Awarded.Date']);
-
-    $numUnlockedSoftcoreProgressions = $userAchievements->where('type', AchievementType::Progression)->whereNotNull('Date')->where('HardcoreMode', UnlockMode::Softcore)->count();
-    $numUnlockedHardcoreProgressions = $userAchievements->where('type', AchievementType::Progression)->whereNotNull('Date')->where('HardcoreMode', UnlockMode::Hardcore)->count();
-    $numUnlockedSoftcoreWinConditions = $userAchievements->where('type', AchievementType::WinCondition)->whereNotNull('Date')->where('HardcoreMode', UnlockMode::Softcore)->count();
-    $numUnlockedHardcoreWinConditions = $userAchievements->where('type', AchievementType::WinCondition)->whereNotNull('Date')->where('HardcoreMode', UnlockMode::Hardcore)->count();
-
-    // If there are no Win Condition achievements in the set, the game is considered beaten
-    // if the user unlocks all the progression achievements.
-    $neededWinConditionAchievements = $totalWinConditions >= 1 ? 1 : 0;
-
-    $isBeatenSoftcore =
-        $numUnlockedSoftcoreProgressions === $totalProgressions
-        && $numUnlockedSoftcoreWinConditions >= $neededWinConditionAchievements;
-
-    $isBeatenHardcore =
-        $numUnlockedHardcoreProgressions === $totalProgressions
-        && $numUnlockedHardcoreWinConditions >= $neededWinConditionAchievements;
-
-    // TODO use $playerGame->beaten_at for isBeatenSoftcore, $playerGame->beaten_hardcore_at for isBeatenHardcore, remove rest
-    return [
-        'isBeatenSoftcore' => $isBeatenSoftcore,
-        'isBeatenHardcore' => $isBeatenHardcore,
-        'isBeatable' => true,
-    ];
-}
 
 function getGameRankAndScore(int $gameID, User $user): array
 {
@@ -340,120 +266,6 @@ function getUsersGameList(User $user): array
     }
 
     return $dataOut;
-}
-
-/**
- * @deprecated TODO: Remove when denormalized data is ready. See comments in getUsersCompletedGamesAndMax().
- */
-function getLightweightUsersCompletedGamesAndMax(string $user, string $cachedAwardedValues): array
-{
-    // Parse the cached value.
-    $awardedCache = [];
-    foreach (explode(',', $cachedAwardedValues) as $row) {
-        list($gameId, $maxPossible, $numAwarded, $numAwardedHC, $mostRecentWonDate, $firstWonDate) = explode('|', $row);
-
-        $awardedCache[$gameId] = [
-            'MaxPossible' => $maxPossible,
-            'NumAwarded' => $numAwarded,
-            'NumAwardedHC' => $numAwardedHC,
-            'MostRecentWonDate' => $mostRecentWonDate,
-            'FirstWonDate' => $firstWonDate,
-        ];
-    }
-
-    $lightQuery = "SELECT gd.ID AS GameID, c.Name AS ConsoleName, c.ID AS ConsoleID, gd.ImageIcon, gd.Title
-    FROM GameData AS gd
-    LEFT JOIN Console AS c ON c.ID = gd.ConsoleID
-    WHERE gd.ID IN (
-        SELECT DISTINCT Achievements.GameID
-        FROM Awarded
-        INNER JOIN Achievements ON Awarded.AchievementID = Achievements.ID
-        WHERE Awarded.User = '$user' AND Achievements.Flags = 3
-    )
-    ORDER BY gd.Title";
-
-    $lightResults = legacyDbFetchAll($lightQuery)->toArray();
-
-    // Merge cached award data
-    foreach ($lightResults as &$game) {
-        $gameId = $game['GameID'];
-
-        $game['MaxPossible'] ??= 0;
-        $game['NumAwarded'] ??= 0;
-        $game['NumAwardedHC'] ??= 0;
-        $game['PctWon'] ??= 0;
-        $game['PctWonHC'] ??= 0;
-
-        if (isset($awardedCache[$gameId])) {
-            $numAwarded = (int) $awardedCache[$gameId]['NumAwarded'];
-            $numAwardedHC = (int) $awardedCache[$gameId]['NumAwardedHC'];
-            $maxPossible = (int) $awardedCache[$gameId]['MaxPossible'];
-            $mostRecentWonDate = $awardedCache[$gameId]['MostRecentWonDate'];
-            $firstWonDate = $awardedCache[$gameId]['FirstWonDate'];
-
-            $game['MaxPossible'] = $maxPossible;
-            $game['NumAwarded'] = $numAwarded;
-            $game['NumAwardedHC'] = $numAwardedHC;
-            $game['PctWon'] = $maxPossible ? $numAwarded / $maxPossible : 0;
-            $game['PctWonHC'] = $maxPossible ? $numAwardedHC / $maxPossible : 0;
-            $game['MostRecentWonDate'] = $mostRecentWonDate;
-            $game['FirstWonDate'] = $firstWonDate;
-        }
-    }
-
-    // Make sure we're sorting correctly similar to the costly query in getUsersCompletedGamesAndMax().
-    usort($lightResults, function ($a, $b) {
-        // Check if either game has 100% achievements won.
-        $a100Pct = (isset($a['PctWon']) && $a['PctWon'] == 1.0);
-        $b100Pct = (isset($b['PctWon']) && $b['PctWon'] == 1.0);
-
-        // If one game has 100% and the other doesn't, sort accordingly.
-        if ($a100Pct && !$b100Pct) {
-            return -1;
-        }
-        if (!$a100Pct && $b100Pct) {
-            return 1;
-        }
-
-        if ($a['PctWon'] != $b['PctWon']) {
-            return $b['PctWon'] <=> $a['PctWon']; // Sort by PctWon descending
-        }
-        if ($a['PctWonHC'] != $b['PctWonHC']) {
-            return $b['PctWonHC'] <=> $a['PctWonHC']; // Sort by PctWonHC descending
-        }
-        if ($a['MaxPossible'] != $b['MaxPossible']) {
-            return $b['MaxPossible'] <=> $a['MaxPossible']; // Sort by MaxPossible descending
-        }
-
-        return $a['Title'] <=> $b['Title']; // Sort by Title ascending
-    });
-
-    // Return combined results
-    return $lightResults;
-}
-
-/**
- * @deprecated TODO Remove when denormalized data is ready. See comments in getUsersCompletedGamesAndMax().
- */
-function prepareUserCompletedGamesCacheValue(array $allFetchedResults): string
-{
-    // Extract awarded data
-    $awardedCacheString = '';
-    foreach ($allFetchedResults as $result) {
-        $gameId = $result['GameID'];
-        $maxPossible = $result['MaxPossible'];
-        $numAwarded = $result['NumAwarded'];
-        $numAwardedHC = $result['NumAwardedHC'];
-        $mostRecentWonDate = $result['MostRecentWonDate'];
-        $firstWonDate = $result['FirstWonDate'];
-
-        $awardedCacheString .= "$gameId|$maxPossible|$numAwarded|$numAwardedHC|$mostRecentWonDate|$firstWonDate,";
-    }
-
-    // Remove last comma
-    $awardedCacheString = rtrim($awardedCacheString, ',');
-
-    return $awardedCacheString;
 }
 
 function getUsersCompletedGamesAndMax(string $user): array
