+++ app/Platform/Actions/ResetPlayerProgress.php
@@ -56,24 +56,17 @@ class ResetPlayerProgress
                     ->delete();
             }
             $playerAchievement->delete();
-
-            $user->playerAchievementsLegacy()->where('AchievementID', $achievementID)->delete();
         } elseif ($gameID !== null) {
             $achievementIds = Achievement::where('GameID', $gameID)->pluck('ID');
 
             $user->playerAchievements()
                 ->whereIn('achievement_id', $achievementIds)
                 ->delete();
-
-            $user->playerAchievementsLegacy()
-                ->whereIn('AchievementID', $achievementIds)
-                ->delete();
         } else {
             // fulfill deletion request
             $user->playerGames()->forceDelete();
             $user->playerBadges()->delete();
             $user->playerAchievements()->delete();
-            $user->playerAchievementsLegacy()->delete();
 
             $user->RAPoints = 0;
             $user->RASoftcorePoints = null;
