+++ app/Platform/AppServiceProvider.php
@@ -42,7 +42,7 @@ use App\Platform\Models\IntegrationRelease;
 use App\Platform\Models\Leaderboard;
 use App\Platform\Models\LeaderboardEntryLegacy;
 use App\Platform\Models\MemoryNote;
-use App\Platform\Models\PlayerAchievementLegacy;
+use App\Platform\Models\PlayerAchievement;
 use App\Platform\Models\PlayerBadge;
 use App\Platform\Models\PlayerBadgeStage;
 use App\Platform\Models\PlayerSession;
@@ -124,8 +124,7 @@ class AppServiceProvider extends ServiceProvider
             'memory-note' => MemoryNote::class,
             'player.badge' => PlayerBadge::class,
             'player.badge-stage' => PlayerBadgeStage::class,
-            // TODO 'player.achievement' => PlayerAchievement::class,
-            'player.achievement' => PlayerAchievementLegacy::class,
+            'player.achievement' => PlayerAchievement::class,
             'player-session' => PlayerSession::class,
             'system' => System::class,
         ]);
