+++ public/userInfo.php
@@ -650,7 +650,7 @@ RenderContentStart($userPage);
 
     if ($user !== null && $user === $userPage) {
         // FIXME: https://discord.com/channels/476211979464343552/1026595325038833725/1162746245996093450
-        // RenderScoreLeaderboardComponent($user, true);
+        // RenderPointsRankingComponent($user, true);
     }
     ?>
 </aside>
